data generated only if not exist

+++ src/app/app.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { DexieService } from './services/dexie.service';
 import {TableQueryService} from './services/table-query.service';
 @Component({
   selector: 'app-root',
@@ -9,13 +10,18 @@ export class AppComponent {
 
   worker = new Worker('./workers/generation.worker', { type: 'module' });
 
-  constructor(public tableQuery: TableQueryService) {
-      const worker = new Worker('./workers/generation.worker', { type: 'module' });
-      worker.onmessage = ({ data }) => {
-        this.tableQuery.bulkAdd(data);
-        worker.terminate();
-      };
-      worker.postMessage({rows: 6000, cols: 100});
+  constructor(public tableQuery: TableQueryService, public dexie: DexieService) {
+    this.dexie.doesExist().then(exists => {
+      if(!exists) {
+        const worker = new Worker('./workers/generation.worker', { type: 'module' });
+        worker.onmessage = ({ data }) => {
+          this.tableQuery.bulkAdd(data);
+          worker.terminate();
+        };
+        worker.postMessage({rows: 6000, cols: 100});
+      }
+    })
+      
     
   }
